[WIP] Top-level await integration
The purpose of this patch is to nail down how HTML would
adjust to the needs of top-level await. The relevant change is,
the module.Evaluate() method will return a Promise that resolves
when the module is ready, and this will be passed to
FinishDynamicImport to wait on before resolving the Promise it returns.
It doesn't seem like anyone else needs to wait for a module script
to "finish" evaluating besides dynamic import.

This PR is just a sketch, not to be landed. Top-level await should
reach stage 3 before this change lands.

The PR where Evaluate returns a Promise is
https://github.com/guybedford/proposal-top-level-await/pull/1